delete pseudo-doc.doxygen (#6536)
We only added that for people who use doxygen to browse code bases. That
was before the age of LSP support in editors, which does a much better
job of that nowadays.

Distributions like Fedora have for some reason started shipping the full
doxygen-generated HTML, which is many megabytes in size. This was never
our intent, so remove the file to prevent it.

fixes https://github.com/i3/i3/issues/6436